html/canvas/element/manual/draw-element-image/privacy/visited-link-slotted-color-ignored.tentative.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not reveal visited link colors</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src="/html/canvas/resources/wait-for-canvas-paint.js"></script>
  <style>
    canvas {
      background-color: blue;
    }
    #link {
      width: 100px;
      height: 100px;
      background-color: green;
      color: green;
      display: block;
    }
    #link:visited {
      background-color: red; /* Visited color */
      color: red; /* Visited color */
    }
  </style>
</head>
<body>
  <canvas id="canvas" width="100" height="100" layoutsubtree>
    <div id="shadowHost">
      <template shadowrootmode="open">
        <div> <!-- add a node so that the shadowHost is not the parent of the link -->
          <slot name="slot1"></slot>
        </div>
      </template>
      <div slot="slot1">
        <a id="link" href="">link</a>
      </div>
    </div>
  </canvas>

  <script>
    window.onload = () => {
      promise_test(async function(t) {
        await waitForCanvasPaint(canvas);

        var context = canvas.getContext("2d");

        // Draw the element image
        context.drawElementImage(shadowHost, 0, 0);

        // The canvas should NOT reflect the visited state (red), it should paint the unvisited state (green).
        let pixel = context.getImageData(10, 10, 1, 1).data;
        assert_false(pixel[0] > 0, "The canvas should not reflect the visited state.");
        assert_true(pixel[1] > 0, "The canvas should paint the slotted link.");
        assert_false(pixel[2] > 0, "The canvas should paint the slotted link.");
      });
    };
  </script>
</body>
</html>
